pyproject.toml
[project]
name = "mcp-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faiss-cpu>=1.10.0",
    "mcp[cli]>=1.6.0",
    "torch>=2.6.0",
]
